+++ src/arch/x86_64/memory/frame/mod.rs
@@ -2,19 +2,19 @@ use memory::PAGE_SIZE;
 
 /// A representation of a physical memory frame
 pub struct Frame {
-    pub number: usize,
+    pub number: u64,
 }
 
 impl Frame {
-    pub fn from_number(number: usize) -> Frame {
+    pub fn from_number(number: u64) -> Frame {
         Frame { number: number }
     }
 
-    pub fn from_address(address: usize) -> Frame {
+    pub fn from_address(address: u64) -> Frame {
         Frame { number: address / PAGE_SIZE }
     }
 
-    pub fn to_address(&self) -> usize {
+    pub fn to_address(&self) -> u64 {
         self.number * PAGE_SIZE
     }
 }
